requirements.txt
faker==22.0.0
pandas==2.1.4
google-cloud-bigquery==3.14.1
pyarrow==14.0.2
db-dtypes==1.2.0
google-cloud==0.34.0
google-cloud-storage==2.14.0